feat(processes): analyse cluster as task
Since user task is not yet available in the Zeebe modeller, let's add a
service task with type humanTask first. Tasklist can pick this up as a
task to work on.

The previously existing message intermediately catch event has become a
boundary event on the task, so stay backwards compatible.

Resolves #205