fix(processes): propagate test results correctly
Previous the test restuls and especially the failure messages haven't
been propagated correctly for two reasons.

 1. there was no output mapping on the call activities
 2. error events terminate the call activities and can't transfer
    payload

In order to overcome this we need to change the process model such that
it a) maps the payload outside the call activity and b) not terminate
the call activity

This commit does that and introduces new variables to verify whether the
call was successful or not.