fix(chaos): replace assembly with shade plugin
In order to fix our Log4J issue, which was caused by the assembly plugin
and building a jar with dependencies, we need to replace the assembly
plugin with the shade plugin and use a third-party plugin to make sure
that the log4j plugin dat file is correctly merged with all
dependencies.

This makes it possible to have the strackdriver logging and building a
jar with dependencies.