fix(chaos): print system error together with system out
Previously the output streams have been printed after another, which can
cause confusion and makes it not really visible what is happening.

Furthemore the worker thread was blocked, until the process was done.
This cause problems on handling jobs, not terminating the process after
some time and we are not seeing what is going on.