fix(chaos-workers): await deploy v1 in multiversion
It could happen that deployment of the initial version of the
multiVersion process would fail. The handler awaits deployment of the
successive versions and in the end expects that 10 different versions
were deployed, but this could be lower if one of the model_v1
deployments failed.

We can mitigate this by also awaiting this deployment until successful.