refactor(cloud-client): isolate ClusterInfo schema
ClusterInfo schema is the response for getClusterInfo and
listClusterInfo. It might have properties that resemble entities
provided by other api calls, but their schemas might differ. To make the
api more resilient against changes to individual endpoints, this commit
isolates the schema.

Make all classes that are used as properties of ChannelInfo as inner
classes of ChannelInfo.