fix(core): split client and cluster creation
This is a workaround for the zeebe url not immediately containing the
region. There is a short delay until the information is updated in
console. Testbench can simply wait a second before retrieving the info.
When this goes wrong, it does not have to delete the cluster (it is
only related to the client).

Note that this workaround can be removed when cluster preprovisioning is
removed with camunda-cloud/console#2225