feat(core): add stackdriver logging
Similar to #363 this uses zeebe-util to reuse the stackdriver layout.

The log4j config defines a Console appender (which is used by default)
and a Stackdriver appender (which can be enabled using ZCTB_LOG_APPENDER
env var). This env var is set by the pod specs.

Regardless of the appender, all info logs are logged, and additionally
the cloud filter is logged at debug (like before).

I've selected `testbench` as the service name as it is short and
understandible. However, we could also consider `ZCTB`, or
`zeebe-cluster-testbench`.

The service version is extracted from the env var
ZCTB_LOG_STACKDRIVER_SERVICEVERSION which is filled by the pod spec
using the namespace. E.g. `testbench-1-x-prod`. This seemed like a
reasonable version name as we use this as well for benchmarks.