build: combine spotless config
The config was defined in 2 places (the plugin management, and the
root-project's build section). By moving the config from the plugin
management section into the build section, it becomes more clear what
the actual config is.

This also helps achieve that the plugin management section only defines
plugin versions, and nothing else.