fix: cleanup mdc once job handling finished
The MDC should be cleaned-up, because otherwise it will continue to use
the added context information in subsequent logging. This adds a
JobHandler decorator that enriches the MDC and cleans it again around
the handling of each activated job.

This approach means that the handlers themselves don't need to know
about this anymore, making the enrichment simpler to specify. We simply
decorate the existing handlers.

The new LoggingEnricher does not use MDC.clear to clean up, but
explicitly cleans all the parts it added. This way it doesn't remove MDC
parts that might have been added by other classes. Note that we don't
have any at this time, AFAIK. But this approach feels a bit safer.