build: use maven-shade-plugin to assemble
An issue exists with log4j in combination with the maven-assembly-plugin
which was keeping us from using the Stackdriver layout. The
maven-shade-plugin allows to resolve this issue by adding a
third-party plugin. This plugin makes sure that the log4j plugin dat
file is correctly merged in the final package.

A similar fix was provided in #365 with a18098238.

This fix was suggested in
https://stackoverflow.com/questions/25065580/log4j2-custom-plugins-annotation-processing-with-maven-assembly-plugin