test(cloud-client): add test for marshalling roundtrip in RestEasy
This test uses wiremock to set up an HTTP server for the unit test.
With this test server it is possible to test a request/response roundtrip
that uses RestEasy marshalling. This is more realistic than using Jackon
directly.

Also while working on this test, I realized that the OAuthCredentials
object could be simplified and did so. In doing so the field "expiry"
was removed. This field was not used in the code. Basically, we try
to use the token, and if it has expired, we will request a new token.