build: append jaxrs Providers in shaded jar
The Providers file located in META-INF/services is supposed to define
the available provider classes. This is where it will look for the
writers. It seemed that the shade plugin somehow merged these files
differently. By defining an appending strategy the file should contain
all entries.

This was an idea from: https://stackoverflow.com/a/18538803/3442860

Note that we use newer version that is not in the javax.ws.rs but in
jakarta.ws.rs namespace.