fix: handle invalid generations
This a shortcut fix or workaround until https://github.com/camunda-cloud/camunda-cloud-management-apps/issues/1731 is fixed
Normally we would expect that we can clone all kind of generations but currently this is not possible.
If we reference a generation without the upgradeableFrom reference (so no reference) we retrieve errors (not found generation) which are hard to debug.
See related issue https://github.com/zeebe-io/zeebe-cluster-testbench/issues/945