ci: dont specify specific maven version
It's unclear to me why these jobs specify a specific maven version. It's
not managed by any version manager, so it quickly gets out of date.

It was introduced with https://github.com/zeebe-io/zeebe-cluster-testbench/pull/802

Instead, it's fine to use the latest maven version for our CI. This is
needed because the flatten maven plugin v1.6 requires a newer maven
version then the one we currently specify in our CI.