fix: check for failed runs in testResults
After removing aggregatedResults we missed to update condition, which caused us to
notify as success even on runs which have failed.

This commit changes the condition to check the testResults array for any FAILED report.
Add more test to verify that the correct paths are taken, for success and failure.